Start to configure a theme contract
In the current configuration, colors and typographic rules are
all defined in _variables.scss. The SCSS module is used to expose
these variables to TypeScript. In order to enforce consistency
across all future themes, an abstract class is used (AbstractTheme)
because interfaces cannot define static fields. The classes folder
exports these files for components. For CSS modules it is necessary
to also add a typings.d.ts file. Since we are using SCSS here, we
furthermore also needed to add the following line inside of it:

declare module "*.module.scss";